change server.port from int to short

+++ inc/Server.hpp
@@ -95,7 +95,7 @@ private:
 
 	std::string name;
 	std::string password;
-	unsigned int port;
+	unsigned short port;
 
 	CmdHandler* cmd_handler;
